agents/flor-cto/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-cloud-secret-manager==2.16.4
requests==2.31.0
gunicorn==21.2.0
